Finish styling announcements container
The announcements container is done, for now at least. I was hoping to
make the text continue on for as many lines as possible before going to
ellipsis, but it seems that might not be possible with just css and
html and incorporating js into this project seems like overkill.